Separate block utility functions and constants
