+++ packages/block/utilities/api/getBlockProps.js
@@ -1,0 +1,5 @@
+export const getBlockProps = ({ blockId, height }) => {
+    if (blockId) return { blockId };
+    if (height) return { height };
+    throw Error('No parameters supplied');
+  };
